requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.54.0
cachetools==5.5.2
charset-normalizer==3.4.2
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
dotenv==0.9.9
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.5.1
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==2.19.0
googleapis-common-protos==1.70.0
huggingface-hub==0.33.0
langchain==0.3.26
langchain-anthropic==0.3.15
langchain-community==0.3.26
langchain-core==0.3.66
langchain-google-firestore==0.5.0
langchain-google-genai==2.1.5
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.4.1
marshmallow==3.26.1
more-itertools==10.7.0
numpy==2.3.1
openai==1.90.0
pydantic==2.11.7
pydantic-settings==2.10.0
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
SQLAlchemy==2.0.41
tiktoken==0.9.0
tokenizers==0.21.1
